\name{geco.BvalSummaryByGene}
\alias{geco.BvalSummaryByGene}
\title{geco.BvalSummaryByGene}
\description{Function to calculate the mean DNA methylation level}
\usage{geco.BvalSummaryByGene(bval=NA, 								   feature=NA								  )}
\arguments{
\item{bval}{Beta value matrix}
\item{feature}{Feature annotation table}
}
\value{Data frame giving the mean beta value for each region of each gene
}
\author{Eric Letouze}
\keyword{methods}
