\name{geco.exportAnnotIGV}
\alias{geco.exportAnnotIGV}
\title{geco.exportAnnotIGV}
\description{Function to export sample annotations in IGV format}
\usage{geco.exportAnnotIGV(annot=NA, 							    sample.col="ID", 							    igvfile=NA 							   )}
\arguments{
\item{annot}{Sample annotation table}
\item{sample.col}{Name of column containing sample IDs in annot table}
\item{igvfile}{File to which the output should be written}
}
\value{No value returned
}
\author{Eric Letouze}
\keyword{methods}
